Update S3 CopyObject example to use sourceBucket & sourceKey parameters
A recent update to the AWS SDK for Java v2
(https://github.com/aws/aws-sdk-java-v2/pull/2612) enabled support for
more user-friendly parameters for specifying the source bucket, key, and
version ID for the S3 CopyObject operation. These new parameters are
also automatically URL-encoded by the SDK, as most users would expect.

This commit updates the dependency to the latest version for the S3
module and takes advantage of the new parameters, simplifying the
example code to no longer need to URL encode.

Related:
https://github.com/awsdocs/aws-doc-sdk-examples/issues/740
https://github.com/awsdocs/aws-doc-sdk-examples/pull/759