Add Swift DeleteObjects example (#3472)
* Move Swift TestWaiter into module
This update moves the TestWaiter type defined previously in the
S3 Basics example into the SwiftUtilities module so that it can be
reused by other example tests that need it.

* Add DeleteObjects sample (work in progress)

* Finish Swift DeleteObjects example and tests

* Remove test output

* Remove unneeded end-tag for metadata

* Simplify deleteObjects() handler
Use a map instead of a loop to build the list of objects to delete.

* Further simplify

* Simplifying Swift code is fun!

* Add DeleteObjects to readme for Swift S3 examples

* Integrate suggestions from Liz

* Removed temporary test from DeleteObjects example
I had previously added a test to DeleteObjects' test suite
to check the behavior of XCTest in a certain case. I forgot
to remove it before my last commit. Now it's gone.

Co-authored-by: Irene Smith <73654695+irenepsmith@users.noreply.github.com>