Swift IAM Basics example and tests
The IAM Basics example with tests for its service handler functions. Also
updates for the Swift IAM `README.md` file and the metadata for the functions
used in this example that don't already have SOS examples listed. This helps
to fill in a few gaps.

(There's also a comment correction in another example's tests, which slipped
into this PR and I'm hoping can just go through as is.)