IAM Basics for Swift (#4122)
* More work on IAM scenario example for Swift

* More fixes

* More progress on getting tests for IAM basics

* Remove temporary hack

* Split service handler into per-service classes

* Swift IAM Basics example and tests

The IAM Basics example with tests for its service handler functions. Also
updates for the Swift IAM `README.md` file and the metadata for the functions
used in this example that don't already have SOS examples listed. This helps
to fill in a few gaps.

(There's also a comment correction in another example's tests, which slipped
into this PR and I'm hoping can just go through as is.)

* Fix duplicate SoS tag

Removes accidental stray `snippet-start` tag.

* Remove stuff added by accident

* Apply editorial changes from Liz

Co-authored-by: Eric Shepherd <sheppy@ip-172-31-43-203.us-east-2.compute.internal>