Add new STimer struct to SwiftUtilities
The `STimer` struct is used to create and run timers. The first feature,
included in this PR, adds a function to wait until a given amount of time
expires, optionally printing messages before and/or after the delay.

This will replace the various delay functions that are currently individually
implemented in various examples and tests throughout the Swift SDK examples.