Updated to allow mocking during testing
Refactored the program to use a protocol defining
the functions used in the Amazon IAM API, which is then
defined either as mock functions or functions to call through
to the AWS SDK for Swift, depending on whether testing or not.

Other than that, code is generally somewhat simpler than before.

I'm planning to replace the example in the "Getting started"
chapter of the Swift SDK guide with this one.