Move cpp lambda Dockerfile to Amazon Linux 2023
The current version of Amazon Linux is AL2023, thus it should be used as
the examples rather than AL2, which will be End of Life mid-2025.

For AL2023, we can skip weak dependencies, and by cleaning the dnf cache
we reduce the container image size (as we do not need to store the
downloaded yum repository metadata in the image).

In this case, the installed set of packages is smaller, and the
resulting zip file for the function is around 40% smaller!

AL2 based: 38MB extracted, 10.5MB compressed
AL2023 based: 22MB extracted, 6.6MB compressed

Memory consumption appears to be the same.