Swift: Simplified example for Swift guide intro (#4983)
This is a simplified Amazon S3 "ListBuckets" example for the AWS SDK for
Swift. There is no mocking support, no test suite, and no command line parsing
or other "extra" features. All the code does is use your default credentials
to dump a list of your Amazon S3 buckets to the screen.