Improvements to swiftbuild tool
Handle errors better when parsing the command line in the `swiftbuild` tool.
In addition, added code that checks each project directory for a file named
`test.sh`. If it exists, it is run instead of `swiftc test`, and its output
code signals sucess (0) or failure (any other value). This allows custom test
procedures.

Added a `test.sh` script to the `ListBuckets-Simple` example. This always
returns 0, signaling success, since there's no test for this example. Without
this, `swiftc test` fails, resulting in the test pass signaling a failure when
there isn't one.

As a bonus tweak, fixed a stray out-of-date syntax issue in the S3 basics
example's `Package.swift` file.