C++: Update hello_s3.cpp (#5254)
Update hello_s3.cpp

Adding explicit authentication check. S3 allows anonymous requests. Due to an issue with the c++ sdk, if a user is not authenticated this method will return an unhelpful ‘0 buckets’ instead of a helpful ‘authentication issue’ message.  Since this example is for new users, lets be more helpful.