Swift: Add mocking example (#5260)
This example demonstrates how to use a protocol to mock the AWS SDK for Swift, since Swift lacks the readwrite reflection required for traditional mocking. Passes the test (which is also the exemplar).

There is no build available on the web since this example is not one that corresponds to a particular function, but instead a technique for Swift developers using the SDK.

Once this PR is submitted, I'll start writing the AWS SDK for Swift Developer Guide chapter that will embed these snippets as it explains the mocking technique used here.