[Rust] Improved comments and added SoS tags
Improve the comments and clarify the code for the `if-modified-since` example.
Added more SoS snippet tags so that parts of the code can be used in the error
handling chapter of the developer guide.

Among the code changes are that the names of a couple of variables were
changed in order to make the code clearer to read out of context.