Swift: Update getting started example to build on Mac (#6061)
Update example for getting started to build on Mac

Fixed some minor issues preventing Mac builds of the ListBuckets-Simple S3
example. These changes will need to be replicated across all examples, but
this is being pushed first because it is a dependency of a documentation
update. This example is embedded in fragments inline in the Swift SDK
developer guide's "Get started" page.

*Changes*

* Added the `platforms` block to the `Package.swift` file. This is needed to
  build for Apple platforms, and was removed from all examples by mistake
  during a cleanup pass many months ago.
* Renamed the `main.swift` file to `entry.swift`. This works around a known
  bug in the Swift build chain when the main program uses asynchronous code.