Swift: Add service error handler to example (#6149)
* Swift: Add service error handler to example

Previous example was getting used to show handling
service errors without actually doing so. Now it does.

Also removed unneeded block from other part of the
code.

* Further improvements to example

Improved comments and text output and fixed a syntax
issue in the service error handler.

* Make error handling follow better practices

This update revises the error handler to show how
to sense and report a specific error code, and
improves other aspects of the code to be clearer
for the reader.