Swift: Add example for retry (#6320)
This PR adds a new example demonstrating how to customize retry behavior using
the AWS SDK for Swift. In addition, the existing example for configuring
services is updated to adapt to recent changes in the Swift SDK, as well as to
make other needed improvements (namely, updating the `Package.swift` file to
include a `platforms` block so the code will build on for macOS and iOS).