JavaScript (v3): Remove messaging app
The messaging app is used in the JS dev guide in the subset guidance section.
The examples is using a deprecated version of jquery. It show how to push and
pull from a queue, but doesn't set it up to happen more than once. The example
is not showing anything that's more useful than the topics and queues workflow.