JavaScript (v3): Rekognition - Delete Rekognition PPE app
The rekognition PPE app is not much different from the photo analyzer
cross service example. The only difference is the action used analyze
the photos. The example also used outdated jquery versions, and needed
heavy refactoring. There are no usages of this app in a dev guide so
I'm removing it.