JavaScript (v3): Rekognition - Remove video analyzer
While we don't have other examples of analyzing video, it's a simple
call that's covered by SDK reference docs. This example is using
deprecated versions of jquery, is documented incorrectly, needs
heavy refactoring, and is showcasing bad practices by doing so much
processing on the client side. Deleting this for now as I believe it's
anti-helpful to customers.