JavaScript (v3): Rekognition, DDB - Remove unhelpful examples (#6817)
* JavaScript (v3): Rekognition - Delete Rekognition PPE app

The rekognition PPE app is not much different from the photo analyzer
cross service example. The only difference is the action used analyze
the photos. The example also used outdated jquery versions, and needed
heavy refactoring. There are no usages of this app in a dev guide so
I'm removing it.

* JavaScript (v3): DynamoDB - Remove item tracker cross service example

The item tracker example was an express app showing basic usage of DynamoDB.
It also had deprecated versions of jquery and needed heavy refactoring to be
up to standard. Other examples cover the DynamoDB usage more thoroughly,
and this example isn't being referenced anywhere.

* JavaScript (v3): Rekognition - Remove video analyzer

While we don't have other examples of analyzing video, it's a simple
call that's covered by SDK reference docs. This example is using
deprecated versions of jquery, is documented incorrectly, needs
heavy refactoring, and is showcasing bad practices by doing so much
processing on the client side. Deleting this for now as I believe it's
anti-helpful to customers.