JavaScript (v3): Glue - Make clean up more robust.
deleteDatabase wasn't being handled, leaving the test open to crashing
before clean up was done. This handles that error.

There was also a potential for objects in the S3 Bucket to not be
fully deleted before trying to delete the bucket. I suspect this
was causing issues leading to N stacks being created every run and
not being torn down.