JavaScript (v3): Integration test fixes. (#6853)
* JavaScript (v3): DynamoDB - Fix hanging test in partiql-batch.

* JavaScript (v3): IAM - Adjust timings in iam basics to prevent timeouts.

* JavaScript (v3): SES - Fix integration test issues.

Several issues were found in the SES integration tests:
- Actual emails were being sent (to a bogus email address) instead of throwing the expected errors.
- There was no waiter checking for S3 object deletion, allowing for a bucket deletion to fail.
- An SES quota had been hit, preventing further tests from running.

The first two issues were addressed, which should prevent the third from happening again.

* Revert "Weathertop: Pause JSV3 runs (#6851)"

This reverts commit b3925e37ca5d3523d22f4bff57f643d899b294ed. Integration
tests were paused in order to debug some failures that were causing the
AWS account to be polluted with resources. The tests are cleaned up in
a previous commit and this turns the tests back on.