JavaScript (v3): Kinesis - Replace web example with single action
The previous example was not configured with server-side encryption,
which is the recommended best practice. It was also a web example
that only showcase one SDK call, which we have decided doesn't meet
the bar of a good example. This replaces the web example with a
a single-action nodejs example and provides updated CDK/CFN scripts.