Swift: Add Sign In With Apple example (#6927)
* Swift example using Sign In With Apple for AWS
This adds a new example that uses Sign In With Apple to authenticate
to AWS, then display a list of the Amazon S3 buckets the user has
available. The user interface is presented using SwiftUI, and the
application works on macOS, iOS, and iPadOS.

* Use SimpleKeychain to save local data
This update switches from using the unencrypted SwiftUI
`@AppStorage` to using Auth0's `SimpleKeychain` library to
save the user account information locally.